+++ core/internal/model/entity/role_permission.go
@@ -2,8 +2,8 @@ package entity
 
 // RolePermission defines the role-permission relationship entity
 type RolePermission struct {
-	Id           int `json:"id"            dc:"Relation ID"`
-	RoleId       int `json:"role_id"       dc:"Role ID"`
-	PermissionId int `json:"permission_id" dc:"Permission ID"`
-	CreateTime   int `json:"create_time"   dc:"Creation time"`
+	Id           int64 `json:"id"            dc:"Relation ID"`
+	RoleId       int64 `json:"role_id"       dc:"Role ID"`
+	PermissionId int64 `json:"permission_id" dc:"Permission ID"`
+	CreateTime   int64 `json:"create_time"   dc:"Creation time"`
 }
